tg-bot/requirements.txt
anyio==4.3.0
APScheduler==3.10.4
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
pocketbase==0.9.2
python-dotenv==1.0.1
python-telegram-bot==20.5
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
